+++ triblespace-core/tests/attributes_macro.rs
@@ -1,6 +1,6 @@
 use triblespace_core::blob::ToBlob;
 use triblespace_core::metadata;
-use triblespace_core::metadata::{Describe, MetaDescribe};
+use triblespace_core::metadata::MetaDescribe;
 use triblespace_core::prelude::valueschemas::ShortString;
 use triblespace_core::prelude::{
     attributes, entity, find, pattern, Attribute, Id, MemoryBlobStore, ToValue, Value,